refactor: PlayScreen のスタイルと枠線を分離
